Fix pointer-as-int use to avoid warnings
The dyld_stub_binding_helper value is actually same size as a
void *.  When building LP64, it does not fit in a uint32_t and
cannot be swapped by OSSwapInt32.

Change the variable it's stored in to uintptr_t and then use the
new OSSwapIntPtr macro to swap it properly for both 32 and 64 bits.